rename: 디렉토리 구조 및 파일 이름 수정 - dto가 늘어남에 따라 도메인 내에서 서브 도메인으로 나누었습니다. - 예외 클래스 이름을 도메인에 맞게 변경했습니다.
