+++ src/main/java/com/example/muaring/domain/social/dto/post/MusicPostDTO.java
@@ -1,4 +1,4 @@
-package com.example.muaring.domain.social.dto;
+package com.example.muaring.domain.social.dto.post;
 
 import lombok.*;
