+++ src/main/java/com/example/muaring/domain/social/service/PostService.java
@@ -6,13 +6,12 @@ import com.example.muaring.domain.group.repository.GroupRepository;
 import com.example.muaring.domain.member.entity.Member;
 import com.example.muaring.domain.member.repository.MemberRepository;
 import com.example.muaring.domain.music.dto.MusicHistoryDTO;
-import com.example.muaring.domain.music.dto.MusicResponseDTO;
 import com.example.muaring.domain.music.entity.Music;
 import com.example.muaring.domain.music.exception.MusicErrorCode;
 import com.example.muaring.domain.music.exception.MusicException;
 import com.example.muaring.domain.music.service.MusicService;
-import com.example.muaring.domain.social.dto.MusicPostDTO;
-import com.example.muaring.domain.social.dto.MusicPostRequestDTO;
+import com.example.muaring.domain.social.dto.post.MusicPostDTO;
+import com.example.muaring.domain.social.dto.post.MusicPostRequestDTO;
 import com.example.muaring.domain.social.repository.MusicPostRepository;
 import com.example.muaring.domain.social.entity.MusicPost;
 import jakarta.transaction.Transactional;
